Обновление используемого функционала Blazor до версии 9
