+++ Sources/Application/Application.csproj
@@ -2,12 +2,8 @@
 
   <PropertyGroup>
 	  <NeutralLanguage>en</NeutralLanguage>
-	  <VersionPrefix>1.2.0</VersionPrefix>
+	  <VersionPrefix>1.3.0</VersionPrefix>
   </PropertyGroup>
-
-  <ItemGroup>
-	  <PackageReference Include="Toolbelt.Blazor.HeadElement" Version="7.*" />
-  </ItemGroup>
 
   <ItemGroup>
 	  <ProjectReference Include="..\Infrastructure\Infrastructure.csproj" />
@@ -15,7 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-	  <EmbeddedResource Update="Resources\Pages\Index.resx">
+	  <EmbeddedResource Update="Resources\Components\Pages\Index.resx">
 		  <Generator></Generator>
 	  </EmbeddedResource>
   </ItemGroup>
