+++ Sources/Application/Components/App.razor
@@ -1,24 +1,15 @@
-﻿@page "/"
-
-@namespace Application.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-
-@{
-	Layout = null;
-}
-
-<!DOCTYPE html>
+﻿<!DOCTYPE html>
 <html lang="en">
 <head>
-	<base href="~/" />
 
+	<base href="/" />
 	<meta name="description"
-		  content="">
+		content="">
 	<meta charset="utf-8">
 	<meta name="viewport"
-		  content="width=device-width, initial-scale=1, shrink-to-fit=no">
+	 	  content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
 	<meta http-equiv="x-ua-compatible"
-		  content="ie=edge">
+	  	  content="ie=edge">
 
 	@*https://docs.microsoft.com/ru-ru/aspnet/core/security/blazor/content-security-policy*@
 	@* TODO: Желательно избавиться от политики 'unsafe-eval'. Пока что без неё блокируется выполнение пакета Toolbelt.Blazor.HeadElement *@
@@ -49,47 +40,27 @@
 			'self';
 		upgrade-insecure-requests;">*@
 
-	<link href="Application.styles.css" rel="stylesheet" />
+	<link rel="stylesheet" href="@Assets["Application.styles.css"]" />
 
 	<!-- Google Font: Roboto -->
 	<link href='https://fonts.googleapis.com/css?family=Roboto:400,100,100italic,300,300italic,400italic,700,700italic,900,900italic&subset=latin,latin-ext,cyrillic,cyrillic-ext'
 		  rel='stylesheet'
 		  type='text/css'>
 
-	<!-- FontAwesome -->
-	<script src="https://kit.fontawesome.com/7bb826d7d6.js"
-			crossorigin="anonymous"></script>
-
 	<!-- Main CSS -->
-	<link rel="stylesheet" href="~/css/main.css">
+	<link rel="stylesheet" href="@Assets["css/main.css"]" />
 
 	<!-- Themes -->
-	<link rel="stylesheet" href="~/css/themes.dark.css">
+	<link rel="stylesheet" href="@Assets["css/themes.dark.css"]" />
 
-	<!-- jQuery -->
-	<environment include="Staging,Production">
-		<script src="https://ajax.aspnetcdn.com/ajax/jQuery/jquery-3.5.1.min.js"></script>
-	</environment>
-	<environment include="Development">
-		<script src="~/js/jquery-3.5.1.js"></script>
-		<script src="~/js/jquery-migrate-3.3.0.js"></script>	<!-- TODO: Убрать миграционный файл после устранения всех предупреждений миграции. -->
-	</environment>
+	<ImportMap />
+	<link rel="icon" type="image/x-icon" href="favicon.ico" />
+	<HeadOutlet @rendermode="InteractiveServer" />
 
-	<!-- dense.js (Retina plugin) -->
-	<script src="~/js/dense.js"></script>
-
-	<!-- ScrollReveal -->
-	<script src="https://cdnjs.cloudflare.com/ajax/libs/scrollReveal.js/3.4.0/scrollreveal.min.js"
-			integrity="sha512-AF/YPW4YtJrYvI5GyFPgPrNICMeROqbfg25vV6T0GhJWyQG3hpfIXmAB3cbFhw3Ebx88tQD+dP5p+d3dyYDkrg=="
-			crossorigin="anonymous"></script>
-
-	<!-- Localization -->
-	<script src="_content/Localization/js/localization.js"></script>
 </head>
 <body class="js-enabled">
 	<app>
-		<component type="typeof(App)"
-				   render-mode="ServerPrerendered" />
+		<Routes @rendermode="InteractiveServer" />
 	</app>
 
 	<!-- TODO: Локализовать сообщения об ошибке и страницу ошибки. -->
@@ -106,6 +77,29 @@
 		<a class="dismiss">🗙</a>
 	</div>
 
-	<script src="_framework/blazor.server.js"></script>
+	<script src="_framework/blazor.web.js"></script>
+
+	<!-- FontAwesome -->
+	<script src="https://kit.fontawesome.com/7bb826d7d6.js"
+			crossorigin="anonymous">
+	</script>
+
+	<!-- jQuery -->
+	<environment include="Staging,Production">
+		<script src="https://ajax.aspnetcdn.com/ajax/jQuery/jquery-3.5.1.min.js"></script>
+	</environment>
+	<environment include="Development">
+		<script src="js/jquery-3.5.1.js"></script>
+		<script src="js/jquery-migrate-3.3.0.js"></script>	<!-- TODO: Убрать миграционный файл после устранения всех предупреждений миграции. -->
+	</environment>
+
+	<!-- dense.js (Retina plugin) -->
+	<script src="js/dense.js"></script>
+
+	<!-- ScrollReveal -->
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/scrollReveal.js/3.4.0/scrollreveal.min.js"
+			integrity="sha512-AF/YPW4YtJrYvI5GyFPgPrNICMeROqbfg25vV6T0GhJWyQG3hpfIXmAB3cbFhw3Ebx88tQD+dP5p+d3dyYDkrg=="
+			crossorigin="anonymous">
+	</script>
 </body>
 </html>
