+++ Sources/Application/Components/CultureSwitcher.razor
@@ -2,6 +2,8 @@
 
 @inherits Localization.Components.CultureSwitcherBase
 
+<PageTitle>@_title</PageTitle>
+
 <a href="javascript:"
    title="@Localizer["SwitchLanguage"]"
    @onclick="@ChangeCulture"
