+++ Sources/Application/Components/Pages/Index.razor
@@ -19,8 +19,8 @@
 				<div class="c-header__avatar">
 					<div class="a-header  c-avatar">
 						<img class="c-avatar__img"
-							 src="images/header/avatar.jpg"
-							 alt="@Localizer["AvatarAlt"]">
+						 	 src="images/header/avatar.jpg"
+						 	 alt="@Localizer["AvatarAlt"]">
 					</div>
 				</div>
