+++ Sources/Application/Components/Routes.razor
@@ -1,6 +1,8 @@
 ﻿@inject IStringLocalizer<App> Localizer
 
-<Router AppAssembly="@typeof(App).Assembly">
+@using Application.Components.Layout
+
+<Router AppAssembly="@typeof(Program).Assembly">
 	<Found Context="routeData">
 		<RouteView RouteData="@routeData"
 				   DefaultLayout="@typeof(MainLayout)" />
