+++ Sources/Application/Components/_Imports.razor
@@ -5,8 +5,9 @@
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.JSInterop
+@using static Microsoft.AspNetCore.Components.Web.RenderMode
+@using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Application
-@using Application.Shared
 @using Application.Components
 @using Microsoft.Extensions.Localization
 @using Localization.Components
