+++ Sources/Application/Program.cs
@@ -1,9 +1,8 @@
 ﻿using System.Globalization;
 
+using Application.Components;
 using Application.Data;
 using Application.Infrastructure.JavaScriptModules.Shared;
-
-using Toolbelt.Blazor.Extensions.DependencyInjection;
 
 WebApplicationBuilder builder = WebApplication.CreateBuilder(args);
 
@@ -12,9 +11,9 @@ ConfigureServicesLocalization(builder.Services);
 builder.Services.AddScoped<MainLayoutJavaScriptModule>();
 
 // Add services to the container.
-builder.Services.AddRazorPages();
-builder.Services.AddServerSideBlazor();
-builder.Services.AddHeadElementHelper();
+builder.Services
+	.AddRazorComponents()
+	.AddInteractiveServerComponents();
 
 builder.Services.AddResumeData();
 
@@ -25,29 +24,29 @@ app.UseSessionLocalization();
 // Configure the HTTP request pipeline.
 if (!app.Environment.IsDevelopment())
 {
-	app.UseExceptionHandler("/Error");
+	app.UseExceptionHandler("/Error", createScopeForErrors: true);
 	// The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
 	app.UseHsts();
 }
 
 app.UseHttpsRedirection();
+app.UseAntiforgery();
 
-app.UseStaticFiles();
+app.MapStaticAssets();
 
-app.UseRouting();
-
-app.MapBlazorHub();
-app.MapFallbackToPage("/_Host");
+app
+	.MapRazorComponents<App>()
+	.AddInteractiveServerRenderMode();
 
 await app.RunAsync().ConfigureAwait(false);
 
 static void ConfigureServicesLocalization(IServiceCollection services)
 {
 	CultureInfo defaultCulture = new("en");
-	CultureInfo[] supportedCultures = new[]
-	{
+	CultureInfo[] supportedCultures =
+	[
 		defaultCulture,
 		new CultureInfo("ru")
-	};
+	];
 	services.AddSessionLocalization(supportedCultures, defaultCulture, "Resources");
 }
