+++ Sources/Application/Resources/Components/CultureSwitcher.resx
@@ -117,6 +117,9 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="SiteTitle" xml:space="preserve">
+    <value>{0} {1} - Resume</value>
+  </data>
   <data name="SwitchLanguage" xml:space="preserve">
     <value>Switch language</value>
   </data>
