+++ Sources/Application/Resources/Components/CultureSwitcher.ru.resx
@@ -117,6 +117,9 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="SiteTitle" xml:space="preserve">
+    <value>{0} {1} - Резюме</value>
+  </data>
   <data name="SwitchLanguage" xml:space="preserve">
     <value>Переключить язык</value>
   </data>
